get_set_data/requirements.txt
requests
pandas
mysql-connector-python
sqlalchemy
statsmodels
scikit-learn
numpy